Improved AI and mayor tile improvement behaviour by Martin Gühmann:
..\ctp2_code\gs\gameobj\CityData.cpp
..\ctp2_code\gs\gameobj\CityData.h
Added NeedMoreFood to determine whether the AI should build a food tile
improvement.

..\ctp2_code\ai\CityManagement\governor.cpp
..\ctp2_code\ai\CityManagement\governor.h

- Removed .NET warnings
- Added Calvitex' roads around cities code
- Removed code that prevents the AI and mayors from building undersea tunnels.
- Modified g_cityAstar.FindRoadPath to match the new signature
- Added more arguments to FindBestTileImprovement function, so that it can
  take
the terrain improvements from the other tiles into consideration.
- Food tile improvements are now build if food is needed and not if there is
plenty of food for the city.

..\ctp2_code\ctp2_code\robot\pathing\CityAstar.cpp
..\ctp2_code\ctp2_code\robot\pathing\CityAstar.h

- Added owner argument to FindRoadPath so that the m_owner variable can be set
  the
actual owner.
- Unexplored terrain is now taken into consideration
- Tiles on that no road improvement can be built, because it has non, e.g dead
  or
the according advance is missing, e.g. water.
- Tiles that no one owns are taken into account, but with higher costs so that
tiles are prefered that are owned, but leave the possiblity open to build a
road
when the territoral connection is missing.
- Foreign tiles are ignored unless the foreigner is an ally, maybe a higher
  penalty
then for unowned tiles should be indroduced.
- Linking cities on different continents and building of undersea tunnels is
  now
also possible.
- To calculate the costs for one tile the MoveCosts are weighted with the pw
  costs
this results in good and cheap pathes.

..\ctp2_code\gs\newdb\StartegyRecord.cpp
..\ctp2_code\gs\newdb\StartegyRecord.h
..\ctp2_code\gs\newdb\StrategyRecord.stamp
..\ctp2_code\gs\newdb\startegy.cdb

Exposed to new constant to startegy.txt to calculate whether food is needed:

code:
	// Maximum number of turns required for city growth of one pop that
	// is accepted for the AI and mayors. Default value == 30.
	//	Bit (Int) TurnsAcceptedForOnePop = 30
	Int TurnsAcceptedForOnePop = 30

	// The number of pops before the current pop maximum the AI and mayors
	// stop to consinder building food terrain improvements in the radius
	// of the according city. Default value == 1.
	//	Bit (Int) StopBuildingFoodBeforePopMax = 1
	Int StopBuildingFoodBeforePopMax = 1



Actual those constants should be Bits so that when two startegies are merged
the value of that startegy in that the value is not defined merged into the
other. But unfortunatly Bits don't allow default values, at least not yet.

..\ctp2_code\ai\ctpai.cpp
Set back the exploration resolution to five, there seem still to be a slow
down in the game for no obvious gain.

..\ctp2_code\ctp2_code\robot\pathing\astar.cpp
..\ctp2_code\ctp2_code\robot\pathing\Astar.h
..\ctp2_code\ctp2_code\robot\pathing\unitastar.cpp
..\ctp2_code\ctp2_code\robot\pathing\UnitAstar.h
..\ctp2_code\ctp2_code\robot\pathing\TradeAstar.cpp
..\ctp2_code\ctp2_code\robot\pathing\RobotAstar.h

Restored the last versions of these files as the check whether a path is
blocked or not has to be put into the EntryCost function or into one of the
from there called other functions, that will reduce the additional code and
the calculation time.

..\ctp2_code\gs\newdb\GoalRecord.cpp
..\ctp2_code\gs\newdb\GoalRecord.h
..\ctp2_code\gs\newdb\GoalRecord.stamp
..\ctp2_code\gs\newdb\goal.cdb

Restored backward compartibility by giving the new flags default values.