2004.10.06 Road and food tile improvement AI behaviour update by Martin Gühmann:
..\ctp2_code\gs\gameobj\unitutil.cpp
..\ctp2_code\gs\gameobj\unitutil.h
..\ctp2_code\ai\CityManagement\governor.cpp
..\ctp2_code\ai\CityManagement\governor.h
Removed an error the should terraform logic and added an assert when the
terrain effect is invalid, just wonder why the could be an invalid terrain
effect. And removed the non fully utilization penalty for food improvements of
ring one cities.

..\ctp2_code\robot\pathing\CityAstar.cpp
Updated road path generation, roads now follow the lowest PW costs instead of
the move costs of that terrain. If the given tile has already a road
improvement then reduced costs are used in order to make the path follow the
already existing road. The roads path can now go trough foreign terrain but
the path costs for that are very high so that such paths are only chose if
there is no other route.

..\ctp2_code\gs\dbgen\Datum.cpp
..\ctp2_code\gs\newdb\*Record.stamp
..\ctp2_code\gs\newdb\StrategyRecord.h
..\ctp2_code\gs\newdb\StrategyRecord.cpp
..\ctp2_code\gs\newdb\strategy.cdb
Updated the database generator and exposed the percent of costs for tiles with
roads already there for the road path generation.